move utils folder to src
