+++ packages/components/src/components/DependencyProvider.js
@@ -1,6 +1,6 @@
 /* eslint-disable sonarjs/no-duplicate-string */
 import { Text } from '@asyncapi/generator-react-sdk';
-import { unsupportedFramework, unsupportedLanguage, invalidRole } from '../../utils/ErrorHandling';
+import { unsupportedFramework, unsupportedLanguage, invalidRole } from '../utils/ErrorHandling';
 
 /**
  * @typedef {'python' | 'javascript' | 'dart' | 'java'} Language
